# swagger/openapi.yaml
{
  "openapi":"3.0.1",
  "info":{
    "title":"My e-shop",
    "version":"v0"
  },
  "servers":[
    {
      "url":"http://localhost:8001/",
      "description":"docker container"
    }
  ],
  "paths":{
    "/product":{
      "put":{
        "tags":[
            "product-controller"
        ],
        "operationId":"updateProductById",
        "requestBody":{
          "content":{
            "application/json":{
              "schema":{
                "$ref":"#/components/schemas/ProductRequest"
              }
            }
          },
          "required":true
        },
        "responses":{
          "412":{
            "description":"Precondition Failed",
            "content":{
              "*/*":{
                "schema":{
                  "$ref":"#/components/schemas/ApiResponse"
                }
              }
            }
          },
          "400":{
            "description":"Bad Request",
            "content":{
              "*/*":{
                "schema":{
                  "$ref":"#/components/schemas/ApiResponse"
                }
              }
            }
          },
          "404":{
            "description":"Not Found",
            "content":{
              "*/*":{
                "schema":{
                  "$ref":"#/components/schemas/ApiResponse"
                }
              }
            }
          },
          "200":{
            "description":"OK",
            "content":{
              "*/*":{
                "schema":{
                  "$ref":"#/components/schemas/ApiResponse"
                }
              }
            }
          }
        }
      },
      "post":{
        "tags":[
            "product-controller"
        ],
        "operationId":"addNewProduct",
        "requestBody":{
          "content":{
            "application/json":{
              "schema":{
                "$ref":"#/components/schemas/ProductRequest"
              }
            }
          },
          "required":true
        },
        "responses":{
          "412":{
            "description":"Precondition Failed",
            "content":{
              "*/*":{
                "schema":{
                  "$ref":"#/components/schemas/ApiResponse"
                }
              }
            }
          },
          "400":{
            "description":"Bad Request",
            "content":{
              "*/*":{
                "schema":{
                  "$ref":"#/components/schemas/ApiResponse"
                }
              }
            }
          },
          "404":{
            "description":"Not Found",
            "content":{
              "*/*":{
                "schema":{
                  "$ref":"#/components/schemas/ApiResponse"
                }
              }
            }
          },
          "200":{
            "description":"OK",
            "content":{
              "*/*":{
                "schema":{
                  "$ref":"#/components/schemas/ApiResponse"
                }
              }
            }
          }
        }
      }
    },
    "/order":{
      "post":{
        "tags":[
            "order-controller"
        ],
        "operationId":"addNewOrder",
        "requestBody":{
          "content":{
            "application/json":{
              "schema":{
                "$ref":"#/components/schemas/OrderRequest"
              }
            }
          },
          "required":true
        },
        "responses":{
          "412":{
            "description":"Precondition Failed",
            "content":{
              "*/*":{
                "schema":{
                  "$ref":"#/components/schemas/ApiResponse"
                }
              }
            }
          },
          "400":{
            "description":"Bad Request",
            "content":{
              "*/*":{
                "schema":{
                  "$ref":"#/components/schemas/ApiResponse"
                }
              }
            }
          },
          "404":{
            "description":"Not Found",
            "content":{
              "*/*":{
                "schema":{
                  "$ref":"#/components/schemas/ApiResponse"
                }
              }
            }
          },
          "200":{
            "description":"OK",
            "content":{
              "*/*":{
                "schema":{
                  "$ref":"#/components/schemas/ApiResponse"
                }
              }
            }
          }
        }
      }
    },
    "/product/getByName":{
      "get":{
        "tags":[
            "product-controller"
        ],
        "operationId":"getProductByName",
        "parameters":[
          {
            "name":"name",
            "in":"query",
            "required":true,
            "schema":{
              "type":"string"
            }
          }
        ],
        "responses":{
          "412":{
            "description":"Precondition Failed",
            "content":{
              "*/*":{
                "schema":{
                  "$ref":"#/components/schemas/ApiResponse"
                }
              }
            }
          },
          "400":{
            "description":"Bad Request",
            "content":{
              "*/*":{
                "schema":{
                  "$ref":"#/components/schemas/ApiResponse"
                }
              }
            }
          },
          "404":{
            "description":"Not Found",
            "content":{
              "*/*":{
                "schema":{
                  "$ref":"#/components/schemas/ApiResponse"
                }
              }
            }
          },
          "200":{
            "description":"OK",
            "content":{
              "*/*":{
                "schema":{
                  "$ref":"#/components/schemas/ApiResponse"
                }
              }
            }
          }
        }
      }
    },
    "/product/getAll":{
      "get":{
        "tags":[
            "product-controller"
        ],
        "operationId":"getAllProducts",
        "responses":{
          "412":{
            "description":"Precondition Failed",
            "content":{
              "*/*":{
                "schema":{
                  "$ref":"#/components/schemas/ApiResponse"
                }
              }
            }
          },
          "400":{
            "description":"Bad Request",
            "content":{
              "*/*":{
                "schema":{
                  "$ref":"#/components/schemas/ApiResponse"
                }
              }
            }
          },
          "404":{
            "description":"Not Found",
            "content":{
              "*/*":{
                "schema":{
                  "$ref":"#/components/schemas/ApiResponse"
                }
              }
            }
          },
          "200":{
            "description":"OK",
            "content":{
              "*/*":{
                "schema":{
                  "type":"string"
                }
              }
            }
          }
        }
      }
    },
    "/product/get/{id}":{
      "get":{
        "tags":[
            "product-controller"
        ],
        "operationId":"getProductById",
        "parameters":[
          {
            "name":"id",
            "in":"path",
            "required":true,
            "schema":{
              "type":"integer",
              "format":"int64"
            }
          }
        ],
        "responses":{
          "412":{
            "description":"Precondition Failed",
            "content":{
              "*/*":{
                "schema":{
                  "$ref":"#/components/schemas/ApiResponse"
                }
              }
            }
          },
          "400":{
            "description":"Bad Request",
            "content":{
              "*/*":{
                "schema":{
                  "$ref":"#/components/schemas/ApiResponse"
                }
              }
            }
          },
          "404":{
            "description":"Not Found",
            "content":{
              "*/*":{
                "schema":{
                  "$ref":"#/components/schemas/ApiResponse"
                }
              }
            }
          },
          "200":{
            "description":"OK",
            "content":{
              "*/*":{
                "schema":{
                  "$ref":"#/components/schemas/ApiResponse"
                }
              }
            }
          }
        }
      }
    },
    "/order/pay/{id}":{
      "get":{
        "tags":[
            "order-controller"
        ],
        "operationId":"payOrder",
        "parameters":[
          {
            "name":"id",
            "in":"path",
            "required":true,
            "schema":{
              "type":"integer",
              "format":"int64"
            }
          }
        ],
        "responses":{
          "412":{
            "description":"Precondition Failed",
            "content":{
              "*/*":{
                "schema":{
                  "$ref":"#/components/schemas/ApiResponse"
                }
              }
            }
          },
          "400":{
            "description":"Bad Request",
            "content":{
              "*/*":{
                "schema":{
                  "$ref":"#/components/schemas/ApiResponse"
                }
              }
            }
          },
          "404":{
            "description":"Not Found",
            "content":{
              "*/*":{
                "schema":{
                  "$ref":"#/components/schemas/ApiResponse"
                }
              }
            }
          },
          "200":{
            "description":"OK",
            "content":{
              "*/*":{
                "schema":{
                  "$ref":"#/components/schemas/ApiResponse"
                }
              }
            }
          }
        }
      }
    },
    "/order/getAll":{
      "get":{
        "tags":[
            "order-controller"
        ],
        "operationId":"getAllProducts_1",
        "responses":{
          "412":{
            "description":"Precondition Failed",
            "content":{
              "*/*":{
                "schema":{
                  "$ref":"#/components/schemas/ApiResponse"
                }
              }
            }
          },
          "400":{
            "description":"Bad Request",
            "content":{
              "*/*":{
                "schema":{
                  "$ref":"#/components/schemas/ApiResponse"
                }
              }
            }
          },
          "404":{
            "description":"Not Found",
            "content":{
              "*/*":{
                "schema":{
                  "$ref":"#/components/schemas/ApiResponse"
                }
              }
            }
          },
          "200":{
            "description":"OK",
            "content":{
              "*/*":{
                "schema":{
                  "type":"string"
                }
              }
            }
          }
        }
      }
    },
    "/order/get/{id}":{
      "get":{
        "tags":[
            "order-controller"
        ],
        "operationId":"getOrderById",
        "parameters":[
          {
            "name":"id",
            "in":"path",
            "required":true,
            "schema":{
              "type":"integer",
              "format":"int64"
            }
          }
        ],
        "responses":{
          "412":{
            "description":"Precondition Failed",
            "content":{
              "*/*":{
                "schema":{
                  "$ref":"#/components/schemas/ApiResponse"
                }
              }
            }
          },
          "400":{
            "description":"Bad Request",
            "content":{
              "*/*":{
                "schema":{
                  "$ref":"#/components/schemas/ApiResponse"
                }
              }
            }
          },
          "404":{
            "description":"Not Found",
            "content":{
              "*/*":{
                "schema":{
                  "$ref":"#/components/schemas/ApiResponse"
                }
              }
            }
          },
          "200":{
            "description":"OK",
            "content":{
              "*/*":{
                "schema":{
                  "$ref":"#/components/schemas/ApiResponse"
                }
              }
            }
          }
        }
      }
    },
    "/product/delete/{id}":{
      "delete":{
        "tags":[
            "product-controller"
        ],
        "operationId":"deleteProductById",
        "parameters":[
          {
            "name":"id",
            "in":"path",
            "required":true,
            "schema":{
              "type":"integer",
              "format":"int64"
            }
          }
        ],
        "responses":{
          "412":{
            "description":"Precondition Failed",
            "content":{
              "*/*":{
                "schema":{
                  "$ref":"#/components/schemas/ApiResponse"
                }
              }
            }
          },
          "400":{
            "description":"Bad Request",
            "content":{
              "*/*":{
                "schema":{
                  "$ref":"#/components/schemas/ApiResponse"
                }
              }
            }
          },
          "404":{
            "description":"Not Found",
            "content":{
              "*/*":{
                "schema":{
                  "$ref":"#/components/schemas/ApiResponse"
                }
              }
            }
          },
          "200":{
            "description":"OK",
            "content":{
              "*/*":{
                "schema":{
                  "$ref":"#/components/schemas/ApiResponse"
                }
              }
            }
          }
        }
      }
    },
    "/order/cancel/{id}":{
      "delete":{
        "tags":[
            "order-controller"
        ],
        "operationId":"cancelOrder",
        "parameters":[
          {
            "name":"id",
            "in":"path",
            "required":true,
            "schema":{
              "type":"integer",
              "format":"int64"
            }
          }
        ],
        "responses":{
          "412":{
            "description":"Precondition Failed",
            "content":{
              "*/*":{
                "schema":{
                  "$ref":"#/components/schemas/ApiResponse"
                }
              }
            }
          },
          "400":{
            "description":"Bad Request",
            "content":{
              "*/*":{
                "schema":{
                  "$ref":"#/components/schemas/ApiResponse"
                }
              }
            }
          },
          "404":{
            "description":"Not Found",
            "content":{
              "*/*":{
                "schema":{
                  "$ref":"#/components/schemas/ApiResponse"
                }
              }
            }
          },
          "200":{
            "description":"OK",
            "content":{
              "*/*":{
                "schema":{
                  "$ref":"#/components/schemas/ApiResponse"
                }
              }
            }
          }
        }
      }
    }
  },
  "components":{
    "schemas":{
      "ApiResponse":{
        "type":"object",
        "properties":{
          "status":{
            "type":"string"
          },
          "message":{
            "type":"string"
          }
        }
      },
      "ProductRequest":{
        "type":"object",
        "properties":{
          "name":{
            "type":"string"
          },
          "quantityInStock":{
            "type":"integer",
            "format":"int32"
          },
          "pricePerUnit":{
            "type":"number",
            "format":"double"
          }
        }
      },
      "ItemRequest":{
        "type":"object",
        "properties":{
          "productName":{
            "type":"string"
          },
          "quantity":{
            "type":"integer",
            "format":"int32"
          }
        }
      },
      "OrderRequest":{
        "type":"object",
        "properties":{
          "items":{
            "type":"array",
            "items":{
              "$ref":"#/components/schemas/ItemRequest"
            }
          }
        }
      }
    }
  }
}